refactor(docker): move info and version to queries
