+++ app/setup-tests/setup-handlers/docker.ts
@@ -1,14 +1,17 @@
 import { DefaultBodyType, PathParams, rest } from 'msw';
 
-import {
-  InfoResponse,
-  VersionResponse,
-} from '@/docker/services/system.service';
+import { InfoResponse } from '@/react/docker/proxy/queries/useInfo';
+import { VersionResponse } from '@/react/docker/proxy/queries/useVersion';
 
 export const dockerHandlers = [
   rest.get<DefaultBodyType, PathParams, InfoResponse>(
     '/api/endpoints/:endpointId/docker/info',
-    (req, res, ctx) => res(ctx.json({}))
+    (req, res, ctx) =>
+      res(
+        ctx.json({
+          Plugins: { Authorization: [], Log: [], Network: [], Volume: [] },
+        })
+      )
   ),
   rest.get<DefaultBodyType, PathParams, VersionResponse>(
     '/api/endpoints/:endpointId/docker/version',
